fix: more fixes for spotify to work, moves functions to auth.ts
